made a folder in components call QuestionForm, moved QuestionForm.js to this folder and created a css file, styled the add question
