+++ client/src/Components/QuestionForm/QuestionForm.js
@@ -1,12 +1,9 @@
 import React, { useState } from "react";
 import { useMutation, gql } from "@apollo/client";
+import "./questionForm.css";
 
 const ADD_QUESTION_MUTATION = gql`
-  mutation AddQuestion(
-    $question: String!
-    $answer: String!
-    $category: String!
-  ) {
+  mutation AddQuestion($question: String!, $answer: String!, $category: String!) {
     addQuestion(question: $question, answer: $answer, category: $category) {
       success
       message
@@ -24,9 +21,7 @@ const QuestionForm = () => {
   const [question, setQuestion] = useState("");
   const [answer, setAnswer] = useState("");
   const [category, setCategory] = useState("MongoDB");
-  const [addQuestion, { data, loading, error }] = useMutation(
-    ADD_QUESTION_MUTATION
-  );
+  const [addQuestion, { data, loading, error }] = useMutation(ADD_QUESTION_MUTATION);
   const [showSuccessMessage, setShowSuccessMessage] = useState(false);
   const [showErrorMessage, setShowErrorMessage] = useState(false);
 
@@ -53,35 +48,22 @@ const QuestionForm = () => {
       <form onSubmit={handleSubmit}>
         <div>
           <label>Question:</label>
-          <input
-            type="text"
-            value={question}
-            onChange={(e) => setQuestion(e.target.value)}
-          />
+          <input type="text" value={question} onChange={(e) => setQuestion(e.target.value)} />
         </div>
 
         <div>
           <label>Answer:</label>
-          <input
-            type="text"
-            value={answer}
-            onChange={(e) => setAnswer(e.target.value)}
-          />
+          <input type="text" value={answer} onChange={(e) => setAnswer(e.target.value)} />
         </div>
 
         <div>
           <label>Category:</label>
-          <select
-            value={category}
-            onChange={(e) => setCategory(e.target.value)}
-          >
+          <select value={category} onChange={(e) => setCategory(e.target.value)}>
             <option value="MongoDB">MongoDB</option>
             <option value="Express">Express</option>
             <option value="React">React</option>
             <option value="Node">Node</option>
-            <option value="JavaScript Fundamentals">
-              JavaScript Fundamentals
-            </option>
+            <option value="JavaScript Fundamentals">JavaScript Fundamentals</option>
             <option value="HTML">HTML</option>
             <option value="CSS">CSS</option>
             <option value="Restful APIs">Restful APIs</option>
@@ -91,15 +73,9 @@ const QuestionForm = () => {
 
         <button type="submit">Submit</button>
       </form>
-      {showSuccessMessage && (
-        <div className="success-message">Question added successfully!</div>
-      )}
+      {showSuccessMessage && <div className="success-message">Question added successfully!</div>}
 
-      {showErrorMessage && (
-        <div className="error-message">
-          Error adding question. Please try again.
-        </div>
-      )}
+      {showErrorMessage && <div className="error-message">Error adding question. Please try again.</div>}
     </div>
   );
 };
