+++ client/src/Components/QuestionForm/questionForm.css
@@ -1,0 +1,62 @@
+/* QuestionForm.css */
+form {
+  display: flex;
+  flex-direction: column;
+  max-width: 500px;
+  margin: auto;
+  padding: 20px;
+  border: 1px solid #ccc;
+  border-radius: 8px;
+}
+
+form div {
+  margin-bottom: 15px;
+}
+
+label {
+  margin-bottom: 5px;
+  font-weight: bold;
+}
+
+input,
+select {
+  padding: 8px;
+  border: 1px solid #ddd;
+  border-radius: 4px;
+  width: 100%;
+}
+
+button {
+  padding: 10px 15px;
+  background-color: var(--primary-color);
+  color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  transition: background-color 0.3s;
+}
+
+button:hover {
+  background-color: var(--secondary-color);
+  color: var(--tirtiary-color);
+  border: 2px solid var(--primary-color);
+}
+
+.success-message,
+.error-message {
+  text-align: center;
+  padding: 10px;
+  margin: 20px;
+
+  border-radius: 4px;
+  color: #fff;
+}
+
+.success-message {
+  background-color: var(--secondary-color);
+  color: var(--tirtiary-color);
+}
+
+.error-message {
+  background-color: #f44336;
+}
