+++ client/src/Components/index.js
@@ -1,4 +1,4 @@
-export { default as QuestionForm } from "./QuestionForm";
+export { default as QuestionForm } from "./QuestionForm/QuestionForm";
 export { default as Header } from "./Header/Header";
 export { default as FlipCard } from "./FlipCard/FlipCard";
 export { default as Categories } from "./Categories/Categories";
